+++ functions/index.js
@@ -1,52 +1,69 @@
+// import firebase from "../../gryphonseniorproject/src/firebaseConfig";
+// import { v4 as uuidv4 } from "uuid";
+
+const uuidv4 = require("uuid/v4")
+const firebase = require("./firebaseConfig").firebase
+
+// The Cloud Functions for Firebase SDK to create Cloud Functions and setup triggers.
 const functions = require('firebase-functions');
 
-const admin = require('firebase-admin')
-const nodemailer = require('nodemailer')
-const cors = require('cors')({origin:true})
+// The Firebase Admin SDK to access Cloud Firestore.
+const admin = require('firebase-admin');
 admin.initializeApp();
-
-// **
-// * Here we're using Gmail to send 
-// */
-let transporter = nodemailer.createTransport({
-    service: 'gmail',
-    auth: {
-        user: 'XXXXX',
-        pass: 'XXXXX'
-    }
-});
-
-exports.sendMail = functions.https.onRequest((req, res) => {
-    cors(req, res, () => {
-      
-        // getting dest email by query string
-        const dest = req.query.dest;
-
-        const mailOptions = {
-            from: 'Storify Team <storify.io@gmail.com>', // Something like: Jane Doe <janedoe@gmail.com>
-            to: dest,
-            subject: 'IM A PICKLE!!!', // email subject
-            text: 'Test',
-            html: `<p style="font-size: 16px;">Pickle Riiiiiiiiiiiiiiiick!!</p>
-                <br />
-                <img src="https://images.prod.meredith.com/product/fc8754735c8a9b4aebb786278e7265a5/1538025388228/l/rick-and-morty-pickle-rick-sticker" />
-            ` // email content in HTML
-        };
-  
-        // returning result
-        return transporter.sendMail(mailOptions, (erro, info) => {
-            if(erro){
-                return res.send(erro.toString());
-            }
-            return res.send('Sended');
-        });
-    });    
-});
+const db = firebase.firestore();
 
 
-// // Create and Deploy Your First Cloud Functions
-// // https://firebase.google.com/docs/functions/write-firebase-functions
-//
-// exports.helloWorld = functions.https.onRequest((request, response) => {
-//  response.send("Hello from Firebase!");
+exports.scheduledFunction = functions.pubsub.schedule('every 1 minutes').onRun((context) => {
+    console.log('This will be run every 1 minutes!');
+
+    db.collection("TimerFunction").add({
+        id: uuidv4(),
+        dateCreated: new Date(),
+    })
+        .then(function () {
+            console.log("Document successfully written!");
+        })
+        .catch(function (error) {
+            console.error("Error writing document: ", error);
+        }); 
+
+    return null;
+  });
+
+
+// exports.scheduledFunctionCrontab = functions.pubsub.schedule('5 11 * * *')
+//   .timeZone('America/New_York') // Users can choose timezone - default is America/Los_Angeles
+//   .onRun((context) => {
+//   console.log('This will be run every day at 11:05 AM Eastern!');
+//   return null;
 // });
+
+
+// // Take the text parameter passed to this HTTP endpoint and insert it into 
+// // Cloud Firestore under the path /messages/:documentId/original
+// exports.addMessage = functions.https.onRequest(async (req, res) => {
+//   // Grab the text parameter.
+//   const original = req.query.text;
+//   // Push the new message into Cloud Firestore using the Firebase Admin SDK.
+//   const writeResult = await admin.firestore().collection('messages').add({original: original});
+//   // Send back a message that we've succesfully written the message
+//   res.json({result: `Message with ID: ${writeResult.id} added.`});
+// });
+
+// // Listens for new messages added to /messages/:documentId/original and creates an
+// // uppercase version of the message to /messages/:documentId/uppercase
+// exports.makeUppercase = functions.firestore.document('/messages/{documentId}')
+//     .onCreate((snap, context) => {
+//       // Grab the current value of what was written to Cloud Firestore.
+//       const original = snap.data().original;
+
+//       // Access the parameter `{documentId}` with `context.params`
+//       console.log('Uppercasing', context.params.documentId, original);
+      
+//       const uppercase = original.toUpperCase();
+      
+//       // You must return a Promise when performing asynchronous tasks inside a Functions such as
+//       // writing to Cloud Firestore.
+//       // Setting an 'uppercase' field in Cloud Firestore document returns a Promise.
+//       return snap.ref.set({uppercase}, {merge: true});
+//     });
