+++ functions/package.json
@@ -12,10 +12,9 @@
     "node": "8"
   },
   "dependencies": {
-    "cors": "^2.8.5",
+    "firebase": "^7.15.0",
     "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.1",
-    "nodemailer": "^6.4.8"
+    "firebase-functions": "^3.6.1"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.2.0"
