+++ package.json
@@ -13,6 +13,8 @@
     "dotenv": "^8.2.0",
     "emailjs-com": "^2.4.1",
     "firebase": "^7.14.6",
+    "firebase-admin": "^8.12.1",
+    "firebase-functions": "^3.6.2",
     "firebase-tools": "8.4.2",
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
